refactor: :recycle: Moving tests to path outside src
